# docs/openapi.yaml
openapi: 3.0.3
info:
  title: ODC Track Backend API
  description: Documentation OpenAPI (spec-first) de l'API ODC Track.
  version: 1.0.0
servers:
  - url: http://localhost:3001/api/v1
    description: Local
  - url: https://odc-track-backend.onrender.com/api/v1
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
paths:
...
  /apprenants/me/documents:
    post:
      tags: [Documents]
      summary: Upload document (apprenant)
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [type, situationId, file]
              properties:
                type:
                  type: string
                  enum: [CONTRAT, ATTESTATION, LETTRE_MISSION, CV, AUTRE]
                  example: CV
                situationId:
                  type: string
                  format: uuid
                  example: 33333333-3333-3333-3333-333333333333
                file:
                  type: string
                  format: binary
      responses:
        '201':
          description: Créé
  /situations/{id}/documents:
    get:
      tags: [Documents]
      summary: Lister les documents d'une situation
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /apprenants/{apprenantId}/documents:
    get:
      tags: [Documents]
      summary: Lister les documents d'un apprenant
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: apprenantId
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
  /documents/{id}:
    get:
      tags: [Documents]
      summary: Détail document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
    delete:
      tags: [Documents]
      summary: Supprimer un document
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...